script/requirements.txt
appdirs
requests
gitpython
